lwres: Update a variable type
This should address two AppVeyor/Visual Studio 2019/x64 warnings:
print-lwres.c(197,23): warning C4267: 'function': conversion from
 'size_t' to 'u_int', possible loss of data
print-lwres.c(201,14): warning C4267: 'return': conversion from
 'size_t' to 'unsigned int', possible loss of data

+++ print-lwres.c
@@ -191,7 +191,7 @@ extern const struct tok ns_class2str[];
 
 static unsigned
 lwres_printname(netdissect_options *ndo,
-                size_t l, const u_char *p0)
+                u_int l, const u_char *p0)
 {
 	ND_PRINT(" ");
 	nd_printjn(ndo, p0, l);
